requirements.txt
# RouteWise AI - Dependencies
# Python 3.9+ recommended

# Data & Geospatial
pandas>=2.0.0
numpy>=1.24.0
geopandas>=0.14.0
osmnx>=1.6.0
networkx>=3.0
shapely>=2.0.0

# ML
scikit-learn>=1.3.0

# NLP / LLM
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
openai>=1.0.0

# Web
flask>=3.0.0
flask-cors>=4.0.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
joblib>=1.3.0
